- Moved the notes index to list folder
