+++ app/javascript/src/components/Dashboard/Notes/List/DeleteAlert.jsx
